Export Excel Sheets to CSV Files

This function reads an Excel file, converts each sheet into a data frame, and writes each sheet to a CSV file in the specified output directory.

Usage

write_excel_sheets_to_csv(excel_path, output_dir = "data/")

Arguments

excel_path

A character string specifying the path to the Excel file.

output_dir

A character string specifying the directory where CSV files will be saved. Defaults to "data/".

Value

A list of file paths corresponding to the exported CSV files.

Examples

if (FALSE) { # \dontrun{
# Export CSV files from an Excel file "Diamonds.xlsx" into the "exported/" directory
write_excel_sheets_to_csv("inst/extdata/Diamonds.xlsx", "inst/exported")
} # }